Support hot update
