+++ codec/codec-fury/src/main/java/com/alipay/sofa/rpc/codec/fury/FuryHelper.java
@@ -29,8 +29,8 @@ import com.alipay.sofa.rpc.log.LogCodes;
  * @author lipan
  */
 public class FuryHelper {
-    private final ConcurrentMap<String, Class[]> requestClassCache = new ConcurrentHashMap<>();
-    private final ConcurrentMap<String, Class> responseClassCache = new ConcurrentHashMap<>();
+    private final ConcurrentMap<String, Class[]> requestClassCache  = new ConcurrentHashMap<>();
+    private final ConcurrentMap<String, Class>   responseClassCache = new ConcurrentHashMap<>();
 
     public Class[] getReqClass(String service, String methodName) {
         String key = buildMethodKey(service, methodName);
@@ -45,6 +45,19 @@ public class FuryHelper {
                 e.printStackTrace();
             }
             loadClassToCache(key, clazz, methodName);
+        } else {
+            // Check if the class loader has changed due to hot update
+            String interfaceClass = ConfigUniqueNameGenerator.getInterfaceName(service);
+            ClassLoader currentClassLoader = getClassLoader(interfaceClass);
+            if (!currentClassLoader.equals(reqClass[0].getClassLoader())) {
+                Class clazz = null;
+                try {
+                    clazz = Class.forName(interfaceClass, true, currentClassLoader);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+                loadClassToCache(key, clazz, methodName);
+            }
         }
         return requestClassCache.get(key);
     }
@@ -62,6 +75,19 @@ public class FuryHelper {
                 e.printStackTrace();
             }
             loadClassToCache(key, clazz, methodName);
+        } else {
+            // Check if the class loader has changed due to hot update
+            String interfaceClass = ConfigUniqueNameGenerator.getInterfaceName(service);
+            ClassLoader currentClassLoader = getClassLoader(interfaceClass);
+            if (!currentClassLoader.equals(respClass.getClassLoader())) {
+                Class clazz = null;
+                try {
+                    clazz = Class.forName(interfaceClass, true, currentClassLoader);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+                loadClassToCache(key, clazz, methodName);
+            }
         }
         return responseClassCache.get(key);
     }
@@ -81,7 +107,7 @@ public class FuryHelper {
         }
         if (pbMethod == null) {
             throw new SofaRpcRuntimeException(LogCodes.getLog(LogCodes.ERROR_METHOD_NOT_FOUND, clazz.getName(),
-                    methodName));
+                methodName));
         }
         Class[] parameterTypes = pbMethod.getParameterTypes();
         if (parameterTypes.length == 0) {
