+++ codec/codec-fury/src/main/resources/user.json
@@ -1,3 +1,3 @@
 {
-  "com.alipay.*": true
+  "com.alipay.sofa.rpc.codec.fury.model.Registered.*": true
 }
