--- codec/codec-fury/src/test/java/com/alipay/sofa/rpc/codec/fury/model/NotRegister.java
@@ -1,8 +1,0 @@
-package com.alipay.sofa.rpc.codec.fury.model;
-
-/**
- * @author lipan
- */
-public class NotRegister {
-    public RegisteredClass request;
-}
