+++ codec/codec-fury/src/test/java/com/alipay/sofa/rpc/codec/fury/model/NotRegister/NotRegister.java
@@ -1,0 +1,10 @@
+package com.alipay.sofa.rpc.codec.fury.model.NotRegister;
+
+import com.alipay.sofa.rpc.codec.fury.model.Registered.RegisteredClass;
+
+/**
+ * @author lipan
+ */
+public class NotRegister {
+    public RegisteredClass request;
+}
