+++ codec/codec-fury/src/test/java/com/alipay/sofa/rpc/codec/fury/model/Registered/RegisteredClass.java
@@ -1,0 +1,11 @@
+package com.alipay.sofa.rpc.codec.fury.model.Registered;
+
+import com.alipay.sofa.rpc.codec.fury.model.NotRegister.NotRegister;
+
+/**
+ * @author lipan
+ */
+public class RegisteredClass {
+    public Integer     key;
+    public NotRegister temp;
+}
