--- codec/codec-fury/src/test/java/com/alipay/sofa/rpc/codec/fury/model/RegisteredClass.java
@@ -1,9 +1,0 @@
-package com.alipay.sofa.rpc.codec.fury.model;
-
-/**
- * @author lipan
- */
-public class RegisteredClass {
-    public Integer     key;
-    public NotRegister temp;
-}
